tovplay-backend/requirements.txt
# TovPlay Backend Dependencies
# Last updated: 2025-12-16

# Core Flask
Flask>=3.1.2
flask-cors>=6.0.1
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5
Flask-Limiter>=4.0.0
Werkzeug>=3.1.3

# Database
SQLAlchemy>=2.0.36
psycopg2-binary>=2.9.11

# Authentication & Security
bcrypt>=5.0.0
PyJWT>=2.10.1
cryptography>=46.0.3
ecdsa>=0.19.1

# API & Validation
marshmallow>=4.1.0
marshmallow-sqlalchemy>=1.4.2
email-validator>=2.3.0

# Async & WebSockets
python-socketio>=5.14.3
python-engineio>=4.12.3
eventlet>=0.40.3
simple-websocket>=1.1.0
greenlet>=3.2.4

# HTTP & Networking
requests>=2.32.5
aiohttp>=3.13.2
aiohappyeyeballs>=2.6.1
aiosignal>=1.4.0
urllib3>=2.5.0
certifi>=2025.11.12
charset-normalizer>=3.4.4
idna>=3.11
yarl>=1.22.0
multidict>=6.7.0
frozenlist>=1.8.0
propcache>=0.4.1

# Scheduling
APScheduler>=3.11.1
tzdata>=2025.2
tzlocal>=5.3.1

# Discord Integration
discord.py>=2.6.4
dnspython>=2.8.0

# Monitoring & Metrics
prometheus-client>=0.19.0
prometheus-flask-exporter>=0.23.0
psutil>=7.1.3

# Server
gunicorn>=23.0.0

# Environment & Config
python-dotenv>=1.2.1

# CLI & Formatting
click>=8.3.0
colorama>=0.4.6
rich>=14.2.0
Pygments>=2.19.2
markdown-it-py>=4.0.0
mdurl>=0.1.2

# Utilities
attrs>=25.4.0
bidict>=0.23.1
blinker>=1.9.0
Deprecated>=1.3.1
h11>=0.16.0
itsdangerous>=2.2.0
Jinja2>=3.1.6
limits>=5.6.0
MarkupSafe>=3.0.3
mypy_extensions>=1.1.0
ordered-set>=4.1.0
packaging>=25.0
pathspec>=0.12.1
platformdirs>=4.5.0
pycparser>=2.23
pytokens>=0.3.0
six>=1.17.0
typing_extensions>=4.15.0
wrapt>=2.0.1
wsproto>=1.3.1
cffi>=2.0.0

# Development (optional)
black>=25.11.0

# Redis (for rate limiting)
redis>=7.0.1
